% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_from_template.R
\name{create_from_template}
\alias{create_from_template}
\title{Create a Quarto file based on defined template}
\usage{
create_from_template(file_name = NULL, template = "article")
}
\arguments{
\item{file_name}{A character string specifying the name of the new
Quarto Markdown file.
If \code{NULL}, the function use \verb{"index}" as the default file name.}

\item{template}{A character string specifying the name of the extension.
Default is \code{"article"}, but can use also \code{"prez"}}
}
\value{
This function does not return a value. It performs file operations to set up
the Quarto IOM story template in the current working directory.
}
\description{
This function creates a Quarto file based on defined template
by copying the necessary template files
into the \verb{_extensions} directory and generating
a new Quarto Markdown (\code{.qmd}) file.
}
\details{
The function performs the following steps:
\itemize{
\item Checks if \code{file_name} is provided. If not, it will be index per default.
\item Verifies what \code{template} is selected.
\item Creates the \verb{_extensions} directory if it does not exist.
\item Creates a subdirectory within \verb{_extensions} for the specified extension.
\item Copies the extension files from the package's internal \code{extdata}
directory to \verb{_extensions}.
\item Checks if the extension files were successfully copied.
\item Creates a new Quarto Markdown file based on the template.
\item Opens the new file in the editor.
}
}
\examples{
\dontrun{
create_from_template(template = "prez")
create_from_template(template = "article")
}
}
