% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iomquarto-package.R
\docType{package}
\name{iomquarto-package}
\alias{iomquarto}
\alias{iomquarto-package}
\title{iomquarto: A collection of Quarto Template for IOM}
\description{
A collection of Quarto Template for IOM for Web Article, Presentation, Report & Story.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iom/iomquarto}
  \item \url{https://iom.github.io/iomquarto/}
  \item Report bugs at \url{https://github.com/iom/iomquarto/issues}
}

}
\author{
\strong{Maintainer}: Edouard Legoupil \email{elegoupil@iom.int}

Other contributors:
\itemize{
  \item IOM [copyright holder]
}

}
\keyword{internal}
